Refactor: 리쿠르팅 페이지 컴파운드 패턴 제거
* recruit.tsx에 있는 데이터를 분리시켜 가독성을 높였습니다.
* 재사용성이 적은 컴포넌트의 컴파운드 패턴을 제거하여 가독성을 높였습니다. (Recruit 컴포넌트)
